refactor(airc-cli): split bearer send modules
