+++ crates/airc-cli/src/bearer/mod.rs
@@ -1,0 +1,7 @@
+pub mod cli;
+pub mod commands;
+
+mod gh;
+mod local_bus;
+mod outcome;
+mod rotate;
